Fix inventory reservation logic, correct order status typos, and update schemas

+++ services/order/app/schemas.py
@@ -10,7 +10,7 @@ from .models import OrderStatus
 class OrderItemCreateRequest(BaseModel):
     product_id: uuid.UUID
     quantity: int = Field(gt=0)
-    unit_price: int = Field(gt=0)
+    unit_price: Decimal = Field(gt=0)
 
 class OrderCreateRequest(BaseModel):
     user_id: uuid.UUID
